Allow algorithm identifier to support both ALG_SIG and ALG_PUBKEY
At the moment, all the algorithms definition we support for
AlgorithmIdentifier parsing are either associated with ALG_SIG
or ALG_PUBKEY but not both. Some algs to come (Ed448 and Ed25519)
use the same OID (the key used to reference an algorithm) for
both subject pubkey and signature, hence the need to allow alg
id to support both. That commit just convert the type to a set
of bits.