Pass context to parsing functions to export useful info
At the moment, the offset and length of various components of the
certificate are made available through that mean (tbs, sig, etc)
but it will be improved in the future to extend that list.

Project validates with the following versions of Frama-C and
dependencies.

$ frama-c --version
23.1 (Vanadium)

$ opam list
...
alt-ergo           2.4.0       The Alt-Ergo SMT prover
alt-ergo-lib       2.4.0       The Alt-Ergo SMT prover library
alt-ergo-parsers   2.4.0       The Alt-Ergo SMT prover parser library
...
frama-c            23.1        Platform dedicated to the analysis of source code written in C
...
why3               1.4.0       Why3 environment for deductive program verificati

$ frama-c -wp-detect
[wp] Prover   Alt-Ergo 2.4.0  [alt-ergo|altergo|Alt-Ergo:2.4.0]
[wp] Prover       CVC4 1.7    [cvc4|CVC4:1.7]
[wp] Prover       CVC4 1.7    [cvc4-ce|CVC4:1.7:counterexamples]
[wp] Prover       CVC4 1.7    [cvc4-strings|CVC4:1.7:strings]
[wp] Prover       CVC4 1.7    [cvc4-strings-ce|CVC4:1.7:strings+counterexamples]
[wp] Prover         Z3 4.8.6  [z3|Z3:4.8.6]
[wp] Prover         Z3 4.8.6  [z3-ce|Z3:4.8.6:counterexamples]
[wp] Prover         Z3 4.8.6  [z3-nobv|Z3:4.8.6:noBV]