Remove doubtful semantic check for keyIdentifier in AKI
Section 4.2.1.1 of RFC 5280 has "The keyIdentifier field of the
authorityKeyIdentifier extension MUST be included in all certificates
generated by conforming CAs to facilitate certification path construction.
There is one exception; where a CA distributes its public key in the form
of a "self-signed" certificate, the authority key identifier MAY be omitted."

Previously, we allowed the lack of keyIdentifier in AKI when the issuer
and subject were identical in the certificate, which was considered as
a *basic* (read poor) way of detecting sel-signed certificates. In pratice,
it is not possible to simply detect a certificate is self-signed w/o
a cryptographic check. That behavior resulted in a lot of otherwise valid
certificates to be considered invalid.

Now that we have a ocntext providing all the useful info to take a decision
regarding this, we can remove the check.