Introduce (DANGEROUS) knob to skip parsing of data value for unknown extensions
Each certificate extension is made of an OID and an associated data value
for which we need a specific parsing function to validate the structure
of the data. This means that by default a certificate will be rejected if
a single extension is unknown. This commit defines two configuration
knobs (macro constants) to allow parsing to continue when encoutering
known but unsupported extensions and completely unknown extensions.

Enabling this knobs makes the parser accept 20% more certificates,
currently passing level of acceptance on our set from 78% to 93%.

Please keep in mind that those options - like all other TEMPORARY_*
options - deactivate an expected useful aspect of the parser. In that
specific case, possibly buggy/crafted extensions data values are
skipped WITHOUT being parsed.

At some point, all those TEMPORARY_* options will be turned off by
default to put the parser in a safe by default configuration.

In order to progress inside